refactor(dashboard): restructure and split modules
