+++ sources/dashboard/src/lib/uplot-react/plugins/index.ts
@@ -1,0 +1,1 @@
+export { chartTooltipPlugin } from "./chart-tooltip-plugin";
